==> FinancialIndex_FridayResampling/01_Merging Data From FTSE Index.ipynb <==
## This notebook is used to generate one excel per index

This notebook is needed because the FTSE index could only be obtained in small ranges of data.

Once this notebook is executed an unique file containing all FTSE data will be generated.

import pandas as pd
import numpy as np
from datetime import date, datetime

#### Merging all files containing FTSE data

ftse = pd.read_csv('input\/^FTSE.csv', parse_dates=True)

for count in range(1,19):
    temp = pd.read_csv('input\/^FTSE ('+str(count)+').csv', parse_dates=True)
    ftse = ftse.append(temp)

ftse = ftse.drop_duplicates(subset=None, keep='first', inplace=False)
len(ftse)

ftse = ftse.sort_values(["Date"], ascending = True)

ftse.to_csv(r'output\/CompleteFTSE.csv', index = False)
