==> FinancialIndex_FridayResampling/03_Merging All Indexes in One Dataframe.ipynb <==
## This notebook merges all indexes using the dates as index and generates one file with daily granularity and one file with weekly granularity

import pandas as pd
from datetime import datetime, timedelta

#Daily
v_sp500 = pd.read_csv('input\/S&P500.csv', parse_dates=True)
v_sp500['Date'] = pd.to_datetime(v_sp500['Date'], infer_datetime_format=True)
v_sp500.index = v_sp500['Date'].values
v_sp500 = v_sp500.loc['2000-01-01':'2020-03-18']
v_sp500 = v_sp500.drop(['Date', 'Adj Close', 'High', 'Low', 'Open'], axis=1)
v_sp500 = v_sp500.add_prefix('sp500_')
v_sp500.tail()

#Daily
v_vix = pd.read_csv('input\/VIX.csv', parse_dates=True)
v_vix['Date'] = pd.to_datetime(v_vix['Date'], infer_datetime_format=True)
v_vix.index = v_vix['Date'].values
v_vix = v_vix.loc['2000-01-01':'2020-03-18']
v_vix = v_vix.drop(['Date', 'Adj Close', 'High', 'Low', 'Open'], axis=1)
v_vix = v_vix.add_prefix('vix_')
v_vix.head()

#Daily
v_dji = pd.read_csv('input\/^DJI.csv', parse_dates=True)
v_dji['Date'] = pd.to_datetime(v_dji['Date'], infer_datetime_format=True)
v_dji.index = v_dji['Date'].values
v_dji = v_dji.loc['2000-01-01':'2020-03-18']
v_dji = v_dji.drop(['Date', 'Adj Close', 'High', 'Low', 'Open', 'High', 'Low', 'Open'], axis=1)
v_dji = v_dji.add_prefix('dji_')
v_dji.head()

#Daily
v_ndx = pd.read_csv('input\/^NDX.csv', parse_dates=True)
v_ndx['Date'] = pd.to_datetime(v_ndx['Date'], infer_datetime_format=True)
v_ndx.index = v_ndx['Date'].values
v_ndx = v_ndx.loc['2000-01-01':'2020-03-18']
v_ndx = v_ndx.drop(['Date', 'Adj Close', 'High', 'Low', 'Open'], axis=1)
v_ndx = v_ndx.add_prefix('ndx_')
v_ndx.head()

#Daily
v_n225 = pd.read_csv('input\/N225.csv', parse_dates=True)
v_n225['Date'] = pd.to_datetime(v_n225['Date'], infer_datetime_format=True)
v_n225.index = v_n225['Date'].values
v_n225 = v_n225.loc['2000-01-01':'2020-03-18']
v_n225 = v_n225.drop(['Date', 'Adj Close', 'High', 'Low', 'Open'], axis=1)
v_n225 = v_n225.add_prefix('n225_')
v_n225.head()

#Daily
v_ftse = pd.read_csv('output\/CompleteFTSE.csv', parse_dates=True)
v_ftse['Date'] = pd.to_datetime(v_ftse['Date'], infer_datetime_format=True)
v_ftse.index = v_ftse['Date'].values
v_ftse = v_ftse.loc['2000-01-01':'2020-03-18']
v_ftse = v_ftse.drop(['Date', 'Adj Close', 'High', 'Low', 'Open'], axis=1)
v_ftse = v_ftse.add_prefix('ftse_')
v_ftse.head()

#Daily
v_hsi = pd.read_csv('input\/^HSI.csv', parse_dates=True)
v_hsi['Date'] = pd.to_datetime(v_hsi['Date'], infer_datetime_format=True)
v_hsi.index = v_hsi['Date'].values
v_hsi = v_hsi.loc['2000-01-01':'2020-03-18']
v_hsi = v_hsi.drop(['Date', 'Adj Close', 'High', 'Low', 'Open'], axis=1)
v_hsi = v_hsi.add_prefix('hsi_')
v_hsi.head()

#Daily
v_n100 = pd.read_csv('input\/^N100.csv')
v_n100['Date'] = pd.to_datetime(v_n100['Date'], infer_datetime_format=True)
v_n100.index = v_n100['Date'].values
v_n100 = v_n100.loc['2000-01-01':'2020-03-18']
v_n100 = v_n100.drop(['Date', 'Adj Close', 'High', 'Low', 'Open'], axis=1)
v_n100 = v_n100.add_prefix('n100_')
v_n100.head()

#Monthly
v_emv =  pd.read_excel('input\/EMV_Data.xlsx', error_bad_lines=False) 
Date = v_emv.apply(lambda x: datetime(int(x['Year']),int(x['Month']),28), axis=1)
v_emv['Date'] = Date
v_emv['Date'] = pd.to_datetime(v_emv['Date'], infer_datetime_format=True)
v_emv.index = v_emv['Date'].values
v_emv = v_emv.loc['2000-01-01':'2020-03-18']
v_emv = v_emv[['Overall EMV Tracker']]
v_emv.head()

#Daily
v_infectious = pd.read_csv('input\/All_Infectious_EMV_Data.csv')
Date = v_infectious.apply(lambda x: datetime(int(x['year']),int(x['month']), int(x['day'])), axis=1)
v_infectious['Date'] = Date
v_infectious['Date'] = pd.to_datetime(v_infectious['Date'], infer_datetime_format=True)
v_infectious.index = v_infectious['Date'].values
v_infectious = v_infectious.loc['2000-01-01':'2020-03-18']
v_infectious = v_infectious.drop(['Date', 'day', 'month', 'year'], axis=1)
v_infectious = v_infectious.add_prefix('infectious_')
v_infectious.head()

#Quarterly - might not use
v_financial_stress = pd.read_excel('input\/Financial_Stress.xlsx')
Date = v_financial_stress.apply(lambda x: datetime(int(x['year']),int(x['quarter'])*3, 28), axis=1)
v_financial_stress['Date'] = Date
v_financial_stress['Date'] = pd.to_datetime(v_financial_stress['Date'], infer_datetime_format=True)
v_financial_stress.index = v_financial_stress['Date'].values
v_financial_stress = v_financial_stress.loc['2000-01-01':'2020-03-18']
v_financial_stress = v_financial_stress.drop(['date', 'Date', 'quarter', 'year'], axis=1)
v_financial_stress = v_financial_stress.add_prefix('financial_')
v_financial_stress.head()

#Monthly
v_gpr = pd.read_excel('input\/gpr_web_latest.xlsx')
v_gpr['Date'] = pd.to_datetime(v_gpr['Date'], infer_datetime_format=True)
v_gpr.index = v_gpr['Date'].values
v_gpr = v_gpr.loc['2000-01-01':'2020-03-18']
v_gpr = v_gpr[['GPR']]
v_gpr.head()

#Monthly
v_trade_uncertainty = pd.read_excel('input\/Trade_Uncertainty_Data.xlsx') 
Date = v_trade_uncertainty.apply(lambda x: datetime(int(x['Year']),int(x['Month']),28), axis=1)
v_trade_uncertainty['Date'] = Date
v_trade_uncertainty['Date'] = pd.to_datetime(v_trade_uncertainty['Date'], infer_datetime_format=True)
v_trade_uncertainty.index = v_trade_uncertainty['Date'].values
v_trade_uncertainty = v_trade_uncertainty.loc['2000-01-01':'2020-03-18']
v_trade_uncertainty = v_trade_uncertainty.drop(['Date', 'Year', 'Month'], axis=1)
v_trade_uncertainty = v_trade_uncertainty.add_prefix('trade_')
v_trade_uncertainty.head()

df = pd.concat([v_sp500, v_vix, v_dji, v_ndx, v_n225, v_ftse, v_hsi, v_n100, v_emv, v_infectious, v_gpr, v_trade_uncertainty], axis=1, join='outer')

at_least_one_value_missing = df.apply(lambda x: (len(x)-x.count())>=1, axis=1)

at_least_one_value_missing.value_counts()

df.info()

df.to_csv(r'output\/CompleteIndexes.csv', index = True)

# Generating weekly sampled dataset
df_friday = datetime(2000,1,7)

df_weekly = pd.DataFrame(columns = df.columns)
df_out_weekly = pd.DataFrame(columns = df.columns)


while df_friday <= df.tail(1).index:
    df_weekly = df_weekly.append(df.loc[df_friday])
    df_friday+=timedelta(days=7)


df_weekly.info()

df_weekly.to_csv(r'output\/CompleteWeeklyIndexes.csv', index = True)

df_weekly.shape

df_weekly.describe()
